Move galil choice logic to server install
